the missing files added

+++ src/academy/everyonecodes/java/week2/set2/exercise5/Shoe.java
@@ -1,10 +1,10 @@
 package academy.everyonecodes.java.week2.set2.exercise5;
 
 public class Shoe {
-    private double size;
+    private int size;
     private String color;
 
-    public Shoe(double size, String color){
+    public Shoe(int size, String color){
         this.size = size;
         this.color = color;
     }
